run `reset -y` if you want to delete everything

+++ maldini/management/commands/reset.py
@@ -6,8 +6,15 @@ class Command(BaseCommand):
 
     help = "Reset the database, removing all documents"
 
-    def handle(self, **options):
-        models.Job.objects.all().delete()
-        models.Digest.objects.all().delete()
-        models.FolderMark.objects.all().delete()
-        models.Document.objects.all().delete()
+    def add_arguments(self, parser):
+        parser.add_argument('-y', action='store_true', dest='really')
+
+    def handle(self, really, **options):
+        if really:
+            models.Job.objects.all().delete()
+            models.Digest.objects.all().delete()
+            models.FolderMark.objects.all().delete()
+            models.Document.objects.all().delete()
+
+        else:
+            print("Not deleting anything; pass -y as flag")
